+++ examples/etst7.cc
@@ -14,10 +14,10 @@ using namespace std;
 int tag,type,num,len;
 ContainerType cType;
 
-vector<unsigned long> uVec1,uVec2;
+vector<unsigned int> uVec1,uVec2;
 double dBuf[100];
 int iBuf[100];
-vector<long> lVec;
+vector<int> lVec;
 evioDOMNodeP parent;
